# swagger_server/models/power.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Power(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, on: bool=None):  # noqa: E501
        """Power - a model defined in Swagger

        :param on: The on of this Power.  # noqa: E501
        :type on: bool
        """
        self.swagger_types = {
            'on': bool
        }

        self.attribute_map = {
            'on': 'on'
        }
        self._on = on

    @classmethod
    def from_dict(cls, dikt) -> 'Power':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Power of this Power.  # noqa: E501
        :rtype: Power
        """
        return util.deserialize_model(dikt, cls)

    @property
    def on(self) -> bool:
        """Gets the on of this Power.


        :return: The on of this Power.
        :rtype: bool
        """
        return self._on

    @on.setter
    def on(self, on: bool):
        """Sets the on of this Power.


        :param on: The on of this Power.
        :type on: bool
        """

        self._on = on